refactor: 각 command 를 src/commands/*.ts 으로 확장가능하도록 리팩토링
